cmake/prebuilt: Use a submodule to host prebuilt dependencies
There are several changes in this commit:

Add definePrebuiltProject and update importThirdPartyLibrary

- New function definePrebuiltProject as a ExternalProject wrapper
- Update importThirdPartyLibrary (Binary and HeaderOnly) functions
- THIRD_PARTY_PREBUILT_PATH points to the submodule root
- Add the third-party-prebuilt submodule
- Update azure pipelines to checkout the submodule
- Refactor the build documentation and include prebuilt info
- SleuthKit has a KP of trying to use a bundled SQLite
- The Augeas lenses are bundled with the prebuilds

This change also reorders the third-party project list in
third-party/CMakeLists.txt so that recursive dependency projects are
defined.

This is a large refactor that updates the API usage for importing
third-party libraries, now they are discovered within the submodule vs.
downloaded and extracted.

The logic is ported from the old provision code for Homebrew and Choco.
This includes patch files (hence the update to .gitignore) and a
inline-replace python script to handle CMake/sed/perl oddities.

The ncurses library is introduced as a recursive dependency but it is
not used in the final build.

The helper ./tools/build_prebuilt.sh is helpful on Linux (and MacOS) to
assure a reproducible build.

This commit applies some feedback and most importantly does not create
the build-from-source targets if the generator is Ninja. This is
because some ExternalProject commands use $(MAKE) and this breaks the
Ninja.build.