cmake/prebuilt: Use a submodule to host prebuilt dependencies
There are several changes in this commit:

Add definePrebuiltProject and update importThirdPartyLibrary

- New function definePrebuiltProject as a ExternalProject wrapper
- Update importThirdPartyLibrary (Binary and HeaderOnly) functions
- THIRD_PARTY_PREBUILT_PATH points to the submodule root
- Add the third-party-prebuilt submodule
- Update azure pipelines to checkout the submodule
- Refactor the build documentation and include prebuilt info
- SleuthKit has a KP of trying to use a bundled SQLite
- The Augeas lenses are bundled with the prebuilds

This change also reorders the third-party project list in
third-party/CMakeLists.txt so that recursive dependency projects are
defined.

This is a large refactor that updates the API usage for importing
third-party libraries, now they are discovered within the submodule vs.
downloaded and extracted.

The logic is ported from the old provision code for Homebrew and Choco.
This includes patch files (hence the update to .gitignore) and a
inline-replace python script to handle CMake/sed/perl oddities.

The ncurses library is introduced as a recursive dependency but it is
not used in the final build.

The helper ./tools/build_prebuilt.sh is helpful on Linux (and MacOS) to
assure a reproducible build.

This commit applies some feedback and most importantly does not create
the build-from-source targets if the generator is Ninja. This is
because some ExternalProject commands use $(MAKE) and this breaks the
Ninja.build.

Move third-party prebuilds to formula folder

Add source-based modules inside the "libraries" folder

  CMake: Move CMake-based libraries outside third-party

  CMake: Import libraries based on the current platform

  CMake: Let the user skip local sources when the config option is empty

  CMake: Add source-based dependencies

  CMake: Remove unused CMakeLists.txt file

  CMake: Initialize submodules for referenced source libraries

  CMake: Point the sqlite/smartmontools to the newly renamed forks

  CMake: Alias the Buck-named glibc target

  CMake: Link sqlite against sleuthkit

  CMake: Update the sqlite submodule

  Azure/Linux: Install Flex and Bison

  Azure/Linux: Temporarily disable formatting checks

  CMake: Fix install steps for Augeas lenses

  CMake: Silence all warnings for source-based libraries

  CMake: Do not compile the sleuthkit tools

  Azure: Comment temporary changes, track the thirdparty-submodules branch

  Remove formula for libraries that have been converted

  CMake: Make Augeas import SELinux headers from gnulib

  CMake: Convert the berkeley-db dependency

  Remove CMake-based berkeley-db formula

  CMake: Convert the librpm library

CMake: Fix installation step for Augeas lenses

CMake/macOS: Acquire Flex and Bison from homebrew

CMake: Convert the libpopt library