Change BUILD_TESTING to OSQUERY_BUILD_TESTS
This is to avoid to enable the tests for the thirdparty libraries which
are checked out from source, since BUILD_TESTING is a common variable
used with CTest to enable tests.

Update azure-pipelines.yml and the BUILD.md guide.

Force BUILD_TESTING to be always off due to a problem with glog
that always turn it on.